+++ src/main/java/com/oceanLink/repository/UserRepo.java
@@ -3,6 +3,6 @@ package com.oceanLink.repository;
 import com.oceanLink.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepo extends JpaRepository<User , Integer> {
+public interface UserRepo extends JpaRepository<User , Long> {
 
 }
